Dismantled MainHelper
- Methods moved or deleted
- Created StartupContextCfgFactory
  - It is required by GlassFishMain, however the StartupContextCfg is used
    also from other places with no access to the factory.
  - Basically it is very similar to BootstrapProperties, but that is not visible
    to GlassFishMain.
