+++ nucleus/core/bootstrap/pom.xml
@@ -129,6 +129,15 @@
                     </archive>
                 </configuration>
             </plugin>
+            <plugin>
+                <groupId>org.apache.felix</groupId>
+                <artifactId>maven-bundle-plugin</artifactId>
+                <configuration>
+                    <instructions>
+                        <Bundle-Activator>com.sun.enterprise.glassfish.bootstrap.osgi.GlassFishMainActivator</Bundle-Activator>
+                    </instructions>
+                </configuration>
+            </plugin>
 
             <plugin>
                 <artifactId>maven-dependency-plugin</artifactId>
@@ -138,7 +147,7 @@
                         <goals>
                             <goal>copy-dependencies</goal>
                         </goals>
-                        <phase>generate-test-resources</phase>
+                        <phase>pre-integration-test</phase>
                         <configuration>
                             <includeArtifactIds>org.apache.felix.main,simple-glassfish-api,glassfish-jul-extension</includeArtifactIds>
                             <outputDirectory>${project.build.directory}/test-osgi</outputDirectory>
@@ -146,15 +155,6 @@
                         </configuration>
                     </execution>
                 </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <configuration>
-                    <instructions>
-                        <Bundle-Activator>com.sun.enterprise.glassfish.bootstrap.osgi.GlassFishMainActivator</Bundle-Activator>
-                    </instructions>
-                </configuration>
             </plugin>
         </plugins>
     </build>
