+++ nucleus/core/bootstrap/src/test/java/com/sun/enterprise/glassfish/bootstrap/GlassFishMainIT.java
@@ -17,7 +17,8 @@
 
 package com.sun.enterprise.glassfish.bootstrap;
 
-import com.sun.enterprise.glassfish.bootstrap.cfg.GFBootstrapProperties;
+import com.sun.enterprise.glassfish.bootstrap.cfg.StartupContextCfg;
+import com.sun.enterprise.glassfish.bootstrap.osgi.impl.FelixAdapter;
 import com.sun.enterprise.glassfish.bootstrap.osgi.impl.OsgiPlatform;
 import com.sun.enterprise.glassfish.bootstrap.osgi.impl.OsgiPlatformAdapter;
 
@@ -33,6 +34,7 @@ import static com.sun.enterprise.glassfish.bootstrap.cfg.BootstrapKeys.INSTALL_R
 import static com.sun.enterprise.glassfish.bootstrap.cfg.BootstrapKeys.INSTANCE_ROOT_PROP_NAME;
 import static com.sun.enterprise.glassfish.bootstrap.cfg.BootstrapKeys.PLATFORM_PROPERTY_KEY;
 import static org.hamcrest.CoreMatchers.anyOf;
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.stringContainsInOrder;
@@ -43,18 +45,18 @@ import static org.osgi.framework.Constants.FRAMEWORK_SYSTEMPACKAGES;
 /**
  * Created by kokil on 5/18/17.
  */
-class MainHelperTest {
+class GlassFishMainIT {
 
     @Test
     void createLauncher_Felix() throws Exception {
-        GFBootstrapProperties properties = createDefaultProperties();
-        OsgiPlatformAdapter osgiPlatformAdapter = OsgiPlatformFactory.getOsgiPlatformAdapter(properties);
-        assertNotNull(osgiPlatformAdapter);
+        StartupContextCfg defaultCfg = createDefaultProperties();
+        OsgiPlatformAdapter osgiPlatformAdapter = OsgiPlatformFactory.getOsgiPlatformAdapter(defaultCfg);
+        assertThat(osgiPlatformAdapter, instanceOf(FelixAdapter.class));
         Properties cfg = osgiPlatformAdapter.readPlatformConfiguration();
         assertNotNull(cfg);
-        cfg.putAll(properties.toProperties());
+        cfg.putAll(defaultCfg.toProperties());
 
-        ClassLoader loader = MainHelper.createLauncherCL(new GFBootstrapProperties(cfg),
+        ClassLoader loader = GlassFishMain.createLauncherCL(new StartupContextCfg(cfg),
             ClassLoader.getPlatformClassLoader());
         assertNotNull(loader);
         Class<?> clazz = loader.loadClass("org.osgi.framework.Bundle");
@@ -71,7 +73,7 @@ class MainHelperTest {
     }
 
 
-    private GFBootstrapProperties createDefaultProperties() throws IOException {
+    private StartupContextCfg createDefaultProperties() throws IOException {
         Properties properties = new Properties();
         properties.setProperty(PLATFORM_PROPERTY_KEY, OsgiPlatform.Felix.name());
         Path installRoot = Files.createTempDirectory("FakeGFInstallRoot");
@@ -86,8 +88,7 @@ class MainHelperTest {
         Files.copy(jarFilesDir.resolve("glassfish-jul-extension.jar"), bootstrapDir.resolve("glassfish-jul-extension.jar"));
         Files.copy(jarFilesDir.resolve("org.apache.felix.main.jar"), felixBin.resolve("felix.jar"));
         Files.copy(jarFilesDir.resolve("simple-glassfish-api.jar"), modulesDir.resolve("simple-glassfish-api.jar"));
-        // This is a fake to avoid making this test an integration test - glassfish.jar file is not yet generated by Maven
-        Files.copy(jarFilesDir.resolve("simple-glassfish-api.jar"), modulesDir.resolve("glassfish.jar"));
+        Files.copy(jarFilesDir.resolve(Path.of("..", "glassfish.jar")), modulesDir.resolve("glassfish.jar"));
 
         Path cfgDir = Files.createDirectories(installRoot.resolve(Path.of("config")));
         Files.createFile(cfgDir.resolve("osgi.properties"));
@@ -95,7 +96,7 @@ class MainHelperTest {
 
         Path instanceRoot = Files.createTempDirectory("FakeGFInstanceRoot");
         properties.setProperty(INSTANCE_ROOT_PROP_NAME, instanceRoot.toFile().getAbsolutePath());
-        return new GFBootstrapProperties(properties);
+        return new StartupContextCfg(properties);
     }
